refactor: extract RancherMeta components into shared pkg/common
Move RancherMeta.vue/RancherMetaProperty.vue and the MetadataProperty
type out of pkg/vulnerability-scanner into a new pkg/common package so
other extensions (e.g. runtime-enforcer) can reuse them without
duplicating source or creating a runtime cross-extension dependency.
