+++ pkg/common/components/RancherMeta.vue
