+++ pkg/common/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "common",
+  "description": "Shared components/types used across security-ui-exts extensions. Not an installable extension itself.",
+  "version": "0.0.0",
+  "private": true
+}
