+++ pkg/common/types.ts
@@ -1,0 +1,7 @@
+export interface MetadataProperty {
+  type: 'text' | 'tags' | 'link' | 'route';
+  label?: string;
+  value?: string;
+  tags?: string[];
+  route: object;
+}
