+++ pkg/vulnerability-scanner/components/ImageDetails.vue
@@ -111,7 +111,7 @@ import { getHighestScore, getPackagePath, getScoreNum, getSeverityNum } from '..
 import DownloadFullReportBtn from './common/DownloadFullReportBtn.vue';
 import DownloadSBOMBtn from './common/DownloadSBOMBtn';
 import MostSevereVulnerabilities from './common/MostSevereVulnerabilities.vue';
-import RancherMeta from './common/RancherMeta.vue';
+import RancherMeta from '@common/components/RancherMeta.vue';
 import VulnerabilityTableSet from './common/VulnerabilityTableSet.vue';
 import WorkloadTableSet from './common/WorkloadTableSet.vue';
