+++ pkg/vulnerability-scanner/components/VexDetails.vue
@@ -56,7 +56,7 @@
 <script>
 import { PRODUCT_NAME, RESOURCE, PAGE, VEX_STATUS } from '@vulnerability-scanner/types';
 import ActionMenu from '@shell/components/ActionMenuShell.vue';
-import RancherMeta from './common/RancherMeta.vue';
+import RancherMeta from '@common/components/RancherMeta.vue';
 import StatusBadge from './common/StatusBadge.vue';
 import { getPermissions } from '@vulnerability-scanner/utils/permissions';
 import day from 'dayjs';
