+++ pkg/vulnerability-scanner/components/__tests__/ImageDetails.spec.ts
@@ -6,7 +6,7 @@ import ImageDetails from '../ImageDetails.vue';
 jest.mock('@components/BadgeState', () => ({ BadgeState: { template: '<div class="badge-state" />' } }));
 jest.mock('../DistributionChart', () => ({ default: { template: '<div class="distribution-chart" />' } }));
 jest.mock('@shell/components/form/LabeledSelect', () => ({ default: { template: '<select class="labeled-select"></select>' } }));
-jest.mock('../common/RancherMeta.vue', () => ({ default: { template: '<div class="rancher-meta" />' } }));
+jest.mock('@common/components/RancherMeta.vue', () => ({ default: { template: '<div class="rancher-meta" />' } }));
 jest.mock('../common/MostSevereVulnerabilities.vue', () => ({ default: { template: '<div class="most-severe" />' } }));
 jest.mock('../common/VulnerabilityTable', () => ({ default: { template: '<div class="vulnerability-table" />' } }));
 jest.mock('../common/DownloadSBOMBtn', () => ({ default: { template: '<button class="download-sbom" />' } }));
