+++ pkg/vulnerability-scanner/types/vulnerability-scanner.ts
@@ -62,10 +62,3 @@ export const PAGE = {
   CVE_DETAIL:      'cve_detail',
   VEX_MANAGEMENT:  'vex_management',
 };
-export interface MetadataProperty {
-  type: 'text' | 'tags' | 'link' | 'route';
-  label?: string;
-  value?: string;
-  tags?: string[];
-  route: object;
-}
