+++ vue.config.js
@@ -15,11 +15,14 @@ module.exports = () => {
       webpackConfig.resolve.alias.set('@vulnerability-scanner', path.resolve(__dirname, 'pkg/vulnerability-scanner'));
       // Add the @runtime-enforcer alias pointing to the pkg/runtime-enforcer folder
       webpackConfig.resolve.alias.set('@runtime-enforcer', path.resolve(__dirname, 'pkg/runtime-enforcer'));
+      // Add the @common alias pointing to the pkg/common folder, shared across extensions
+      webpackConfig.resolve.alias.set('@common', path.resolve(__dirname, 'pkg/common'));
     };
   } else {
     vendorConfig.chainWebpack = (webpackConfig) => {
       webpackConfig.resolve.alias.set('@vulnerability-scanner', path.resolve(__dirname, 'pkg/vulnerability-scanner'));
       webpackConfig.resolve.alias.set('@runtime-enforcer', path.resolve(__dirname, 'pkg/runtime-enforcer'));
+      webpackConfig.resolve.alias.set('@common', path.resolve(__dirname, 'pkg/common'));
     };
   }
